LayoutTests/platform/ios-wk1/fast/loader/window-open-to-invalid-url-disallowed-expected.txt

main frame - willPerformClientRedirectToURL: http://A=a%25B=b
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
main frame - didCancelClientRedirectForFrame
Tests that we do not open a new window to an invalid URL. This test PASSED if you see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: http://A=a%B=b" followed by "didCancelClientRedirectForFrame".

Note, this test must be run in DumpRenderTree.
